# public/assets/api.yaml
openapi: 3.0.0
paths:
  /sysDepts:
    get:
      tags:
        - SysDept
      summary: 'Find list SysDept'
      description: 'Returns list of SysDept'
      operationId: getSysDept
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysDept' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysDept' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysDept' } }
                type: object
        '404':
          description: 'SysDept not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysDept
      summary: 'Add a new SysDept to the store'
      operationId: addSysDept
      requestBody:
        $ref: '#/components/requestBodies/SysDept'
      responses:
        '201':
          description: 'Created SysDept'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDept'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysDept'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysDepts/{id}':
    get:
      tags:
        - SysDept
      summary: 'Find SysDept by ID'
      description: 'Returns a single SysDept'
      operationId: getSysDeptById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysDept to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDept'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysDept'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysDept not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysDept
      summary: 'Update an existing SysDept'
      operationId: updateSysDept
      parameters:
        -
          name: id
          in: path
          description: 'SysDept id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysDept'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysDept not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysDept
      summary: 'Deletes a SysDept'
      operationId: deleteSysDept
      parameters:
        -
          name: id
          in: path
          description: 'SysDept id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /sysDictItems:
    get:
      tags:
        - SysDictItem
      summary: 'Find list SysDictItem'
      description: 'Returns list of SysDictItem'
      operationId: getSysDictItem
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysDictItem' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysDictItem' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysDictItem' } }
                type: object
        '404':
          description: 'SysDictItem not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysDictItem
      summary: 'Add a new SysDictItem to the store'
      operationId: addSysDictItem
      requestBody:
        $ref: '#/components/requestBodies/SysDictItem'
      responses:
        '201':
          description: 'Created SysDictItem'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDictItem'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysDictItem'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysDictItems/{id}':
    get:
      tags:
        - SysDictItem
      summary: 'Find SysDictItem by ID'
      description: 'Returns a single SysDictItem'
      operationId: getSysDictItemById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysDictItem to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDictItem'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysDictItem'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysDictItem not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysDictItem
      summary: 'Update an existing SysDictItem'
      operationId: updateSysDictItem
      parameters:
        -
          name: id
          in: path
          description: 'SysDictItem id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysDictItem'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysDictItem not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysDictItem
      summary: 'Deletes a SysDictItem'
      operationId: deleteSysDictItem
      parameters:
        -
          name: id
          in: path
          description: 'SysDictItem id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
    SysDept:
      title: SysDept
      description: SysDept
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
        name:
          nullable: true
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 64
        parent_id:
          nullable: true
          title: parent_id
          description: parent_id
          type: integer
          format: '-'
          maxLength: 20
        tree_path:
          nullable: true
          title: tree_path
          description: tree_path
          type: string
          format: '-'
          maxLength: 255
        sort:
          nullable: true
          title: sort
          description: sort
          type: integer
          format: '-'
          maxLength: 11
        status:
          nullable: true
          title: status
          description: status
          type: integer
          format: '-'
          maxLength: 1
        deleted:
          nullable: true
          title: deleted
          description: deleted
          type: integer
          format: '-'
          maxLength: 1
        gmt_create:
          nullable: true
          title: gmt_create
          description: gmt_create
          type: string
          format: date
        gmt_modified:
          nullable: true
          title: gmt_modified
          description: gmt_modified
          type: string
          format: date
      type: object
...
    SysDictItem:
      title: SysDictItem
      description: SysDictItem
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
        name:
          nullable: true
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 50
        value:
          nullable: true
          title: value
          description: value
          type: string
          format: '-'
          maxLength: 50
        dict_code:
          nullable: true
          title: dict_code
          description: dict_code
          type: string
          format: '-'
          maxLength: 50
        sort:
          nullable: true
          title: sort
          description: sort
          type: integer
          format: '-'
          maxLength: 11
        status:
          nullable: true
          title: status
          description: status
          type: integer
          format: '-'
          maxLength: 1
        defaulted:
          nullable: true
          title: defaulted
          description: defaulted
          type: integer
          format: '-'
          maxLength: 1
        remark:
          nullable: true
          title: remark
          description: remark
          type: string
          format: '-'
          maxLength: 255
        gmt_create:
          nullable: true
          title: gmt_create
          description: gmt_create
          type: string
          format: date
        gmt_modified:
          nullable: true
          title: gmt_modified
          description: gmt_modified
          type: string
          format: date
      type: object
...
  requestBodies:
    SysDept:
      description: 'SysDept object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysDept'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysDept'
        application/json:
          schema:
            $ref: '#/components/schemas/SysDept'
...
    SysDictItem:
      description: 'SysDictItem object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysDictItem'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysDictItem'
        application/json:
          schema:
            $ref: '#/components/schemas/SysDictItem'
...